Issue-639: Refactor script.js to speed up first section loading
